Refine tests
